agents/CarDataagent/requirements.txt
fastapi
uvicorn
python-dotenv
crewai
crewai-tools
masumi
pydantic
python-multipart
httpx